Clean up tests; however, remove many redundant things like BigDecimalSummer that are no longer needed
