+++ aurora4spark-parent/aurora4spark-sql-plugin/src/main/scala/com/nec/spark/planning/VeoGenericPlanExtractor.scala
@@ -1,4 +1,5 @@
 package com.nec.spark.planning
+
 import com.nec.spark.agile.AggregationExpression
 import com.nec.spark.agile.AggregationFunction
 import com.nec.spark.agile.AttributeName
