--- aurora4spark-parent/aurora4spark-sql-plugin/src/test/scala/com/nec/spark/agile/BigDecimalSummerSpec.scala
@@ -1,15 +1,0 @@
-package com.nec.spark.agile
-
-import com.nec.spark.agile.BigDecimalSummer.ScalaSummer
-import org.scalatest.freespec.AnyFreeSpec
-
-final class BigDecimalSummerSpec extends AnyFreeSpec {
-
-  private def sumTest(summer: BigDecimalSummer): Unit =
-    assert(summer.sum(List(1, 2, 3, 4)) == BigDecimal(10))
-
-  "It works in-JVM" in {
-    sumTest(ScalaSummer)
-  }
-
-}
